requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0

# Tool dependencies
langchain-community>=0.0.10
wikipedia>=1.4.0
arxiv>=1.4.7
duckduckgo-search>=3.9.0
numpy>=1.24.0
sympy>=1.12.0

# Utility dependencies
protobuf>=4.25.1